Rating from the users post-delivery of the orders. Order Delivered notification not popped up. Order Id at user app is NA. Date & time filter in pandit booking should be there. Home page just displaying 6 cards. Previously it was good most trending, Daily Rituals,Popular…..etc.
Currently Pandit booking will capture as per user location but user can also book pandit out station. So there should be filter of location also while selecting the pandit.

Delivery Charge (Order value less than 500),platform charge, COD charge option should be there to manage

+++ lib/model/request/payment_req/payment_reqs_models.dart
@@ -1,5 +1,5 @@
 class CreateOrderReqModel {
-  final int deliveryFee;
+  final double deliveryFee;
   final List<VerifyItem> items;
   final String? couponCode;
   final String? offerId;
